64 bit machine is not happy with converting a pointer to int directory

+++ examples/charm++/piArray/pgm.C
@@ -57,7 +57,7 @@ void main::results(int cnt)
 
 piPart::piPart()
 { 
-  CrnSrand((int) this);
+  CrnSrand((int)(long) this);
 }
 
 void
